requirements.txt
crawl4ai==0.4.24
beautifulsoup4
langchain
langchain-google-genai
google-generativeai
python-dotenv
